+++ NEWS.md
@@ -1,6 +1,10 @@
 *optweight* News and Updates
 ======
 
+# *optweight* (development version)
+
+* Replaced *chk* dependency with *arg*.
+
 # *optweight* 2.0.1
 
 * Added support for the new version of *osqp*, which changes some optional argument names and defaults. These should not impact results.